refactor: 검색결과 화면 페이지 컴포넌트 나누기
